Update to baseType2014.xsd
The definition of Record, and RecordType has been incomplete as per definition in ISO 19103:2015 with the classes Filed and FieldType being completely omitted. Two alternatives: amend ISO 19103 or complicate RecordType (not a simple string).
This update may cause some angst for those users that have been using the existing XSD, consider renaming to baseType2015.xsd (inline with the release of ISO 19103:2015) as this would allow continued use of existing schema.